requirements-test.txt
# Testing dependencies for the trading backend

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1

# Async testing
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Redis for testing
redis==5.0.1
aioredis==2.0.1

# Data validation and serialization
pydantic==2.5.0
marshmallow>=3.20.1

# HTTP client for testing
httpx==0.25.2
requests==2.31.0

# Mocking and faking
faker==20.1.0
factory-boy==3.3.0

# Performance testing
locust==2.17.0
pytest-benchmark==4.0.0

# Security testing
bandit==1.7.5
safety>=2.3.5

# Code quality
black>=23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Documentation testing
doctest-ignore-unicode==0.1.2

# Environment management
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# JSON handling
jsonschema==4.20.0

# Logging for tests
structlog==23.2.0

# Process monitoring (for performance tests)
psutil==5.9.6

# JWT for authentication testing
PyJWT==2.8.0
cryptography==45.0.5

# Database testing (if needed)
sqlalchemy==2.0.23
alembic==1.13.1

# WebSocket testing
websockets==12.0

# Statistical analysis for performance tests
numpy==1.25.2
pandas==2.1.4

# Visualization for test reports
matplotlib==3.8.2
seaborn==0.13.0 